fix(spree): refactor cart handling and improve checkout integration
