+++ platform/canvas-packages/internal_pkgs/spree/src/cart/use-cart.tsx
@@ -3,20 +3,9 @@ import type { SWRHook } from '@plasmicpkgs/commerce'
 import { useCart as useCommerceCart, UseCart } from '@plasmicpkgs/commerce'
 import type { GetCartHook } from '../types/cart'
 import normalizeCart from '../utils/normalizations/normalize-cart'
-import type { GraphQLFetcherResult } from '../types'
-import type { IOrder } from '@spree/storefront-api-v2-sdk/types/interfaces/Order'
-import type { IToken } from '@spree/storefront-api-v2-sdk/types/interfaces/Token'
-import { FetcherError } from '@plasmicpkgs/commerce'
-import { setCartToken } from '../utils/tokens/cart-token'
-import ensureIToken from '../utils/tokens/ensure-itoken'
-import isLoggedIn from '../utils/tokens/is-logged-in'
-import createEmptyCart from '../utils/create-empty-cart'
-import { requireConfigValue } from '../isomorphic-config'
+import getCart from '../utils/get-cart'
 
-const imagesSize = requireConfigValue('imagesSize') as string
-const imagesQuality = requireConfigValue('imagesQuality') as number
-
-export default useCommerceCart as UseCart<typeof handler>;
+export default useCommerceCart as UseCart<typeof handler>
 
 // This handler avoids calling /api/cart.
 // There doesn't seem to be a good reason to call it.
@@ -36,64 +25,7 @@ export const handler: SWRHook<GetCartHook> = {
       options
     )
 
-    let spreeCartResponse: IOrder | null
-
-    const token: IToken | undefined = ensureIToken()
-
-    if (!token) {
-      spreeCartResponse = null
-    } else {
-      try {
-        const { data: spreeCartShowSuccessResponse } = await fetch<
-          GraphQLFetcherResult<IOrder>
-        >({
-          variables: {
-            methodPath: 'cart.show',
-            arguments: [
-              token,
-              {
-                include: [
-                  'line_items',
-                  'line_items.variant',
-                  'line_items.variant.product',
-                  'line_items.variant.product.images',
-                  'line_items.variant.images',
-                  'line_items.variant.option_values',
-                  'line_items.variant.product.option_types',
-                ].join(','),
-                image_transformation: {
-                  quality: imagesQuality,
-                  size: imagesSize,
-                },
-              },
-            ],
-          },
-        })
-
-        spreeCartResponse = spreeCartShowSuccessResponse
-      } catch (fetchCartError) {
-        if (
-          !(fetchCartError instanceof FetcherError) ||
-          fetchCartError.status !== 404
-        ) {
-          throw fetchCartError
-        }
-
-        spreeCartResponse = null
-      }
-    }
-
-    if (!spreeCartResponse || spreeCartResponse?.data.attributes.completed_at) {
-      const { data: spreeCartCreateSuccessResponse } = await createEmptyCart(
-        fetch
-      )
-
-      spreeCartResponse = spreeCartCreateSuccessResponse
-
-      if (!isLoggedIn()) {
-        setCartToken(spreeCartResponse.data.attributes.token)
-      }
-    }
+    const spreeCartResponse = await getCart(fetch)
 
     return normalizeCart(spreeCartResponse, spreeCartResponse.data)
   },
